+++ packages/boot/src/components/app/router.tsx
@@ -94,10 +94,16 @@ function createRouteFromDefinition( route: Route, parentRoute: AnyRoute ) {
 				inspector = await routeConfig.inspector( context );
 			}
 
+			let stage = true;
+			if ( routeConfig.stage ) {
+				stage = await routeConfig.stage( context );
+			}
+
 			return {
 				...( loaderData as any ),
 				canvas: canvasData,
 				inspector,
+				stage,
 				title: titleData,
 				routeContentModule: route.content_module,
 			};
@@ -116,12 +122,12 @@ function createRouteFromDefinition( route: Route, parentRoute: AnyRoute ) {
 
 		return createLazyRoute( route.path )( {
 			component: function RouteComponent() {
-				const { inspector: showInspector } =
+				const { inspector: showInspector, stage: showStage } =
 					useLoaderData( { from: route.path } ) ?? {};
 
 				return (
 					<>
-						{ Stage && (
+						{ Stage && showStage && (
 							<div className="boot-layout__stage">
 								<Stage />
 							</div>